lionpike-contact-fn/requirements.txt
azure-functions
azure-identity
azure-keyvault-secrets
requests